pyproject.toml
[project]
name = "perplexity-2-0-server"
version = "0.1.0"
description = "Backend of perplexity 2.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136.1",
    "langchain>=1.2.18",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-tavily>=0.2.18",
    "langgraph>=1.1.10",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.46.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]
